requirements.txt
aiohttp
aiofiles
python-dotenv
websockets
psutil
numpy
motor
pymongo
qdrant-client
networkx
requests

# Testing dependencies
pytest
pytest-asyncio
pytest-cov
